auth/pyproject.toml
[tool.poetry]
name = "async-architecture-6"
version = "0.1.0"
description = ""
authors = ["Pavel Korsakov <PKorsakov@datafort.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
fastapi-users-db-sqlalchemy = "^6.0.1"
uvicorn = "^0.27.1"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
argon2-cffi = "^23.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
black = "^24.2.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
